fix `this` owner name filtering
