+++ common/src/main/org/jetbrains/lincheck/analysis/ShadowStack.kt
@@ -72,23 +72,3 @@ fun ShadowStackFrame.findLocalVariableFieldReferringTo(obj: Any): OwnerName? {
     }
     return null
 }
-
-/**
- * Checks if a string matches patters of `this`-like local variable or field name.
- *
- * These include:
- * - `this` --- simply `this` local variable name;
- * - `$this`, `this_$iv`--- synthetic local variable names generated to refer to `this` inside inline function;
- * - `this$N` --- field name generated to refer to enclosing `this` object in a nested class;
- *
- * These names should not be used as "owner names" of the method call or field access, as they are "hidden".
- */
-fun isThisName(name: String): Boolean =
-    (name == "this")                                    ||
-    (name == "\$this")                                  ||
-    name.matches(Regex("this\\$\\d+"))                  ||
-    isInlineThisIVName(name)
-
-fun isInlineThisIVName(name: String): Boolean =
-    name.startsWith("this_\$iv")                        ||
-    name.contains(Regex("^\\\$this\\$.+?(\\\$iv)+$"))
