+++ common/src/main/org/jetbrains/lincheck/trace/TraceUtils.kt
@@ -10,6 +10,10 @@
 
 package org.jetbrains.lincheck.trace
 
+import org.jetbrains.lincheck.descriptors.AccessLocation
+import org.jetbrains.lincheck.descriptors.AccessPath
+import org.jetbrains.lincheck.descriptors.FieldAccessLocation
+import org.jetbrains.lincheck.descriptors.LocalVariableAccessLocation
 import org.jetbrains.lincheck.util.isJavaLambdaClass
 
 fun String.adornedClassNameRepresentation(): String = this
@@ -17,6 +21,35 @@ fun String.adornedClassNameRepresentation(): String = this
     .getSimpleClassName()
     .removeJavaLambdaRuntimeAddress()
     .replaceNestedClassDollar()
+
+/**
+ * Checks if a string matches patters of `this`-like local variable or field name.
+ *
+ * These include:
+ * - `this` --- simply `this` local variable name;
+ * - `$this`, `this_$iv`--- synthetic local variable names generated to refer to `this` inside inline function;
+ * - `this$N` --- field name generated to refer to enclosing `this` object in a nested class;
+ *
+ * These names should not be used as "owner names" of the method call or field access, as they are "hidden".
+ */
+fun isThisName(name: String): Boolean =
+    (name == "this")                                    ||
+    (name == "\$this")                                  ||
+    name.matches(Regex("this\\$\\d+"))                  ||
+    isInlineThisIVName(name)
+
+fun isInlineThisIVName(name: String): Boolean =
+    name.startsWith("this_\$iv")                        ||
+    name.contains(Regex("^\\\$this\\$.+?(\\\$iv)+$"))
+
+fun AccessLocation.isThisAccess(): Boolean = when (this) {
+    is LocalVariableAccessLocation  -> isThisName(variableName)
+    is FieldAccessLocation          -> isThisName(fieldName)
+    else                            -> false
+}
+
+fun AccessPath.filterThisAccesses(): AccessPath =
+    AccessPath(locations.filter { !it.isThisAccess() })
 
 /**
  * Replaces nested class dollars (if present) from string with dots.
